requirements.txt
datasketch~=1.6.5
GitPython~=3.1.44
magika
matplotlib~=3.10.1
pandas~=2.2.3
python-magic
py-tlsh
simhash~=2.1.2
spookyhash~=2.1.0
zstandard~=0.23.0
numpy~=1.26.4